# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking — User API",
        "description": "Customer-facing REST API for mobile/web user UI. Authenticate with Bearer token from `POST /user/auth/login` or `/user/auth/register`.\n\n## Test / Dummy Credentials (from seeders)\n\n**All accounts use password:** `password123`\n\n### Customer — API login (`POST /user/auth/login`) & web (`/customer/login`)\n| Name | Email |\n|------|-------|\n| Amit Verma | amit@customer.com |\n| Sneha Patel | sneha@customer.com |\n| Rahul Singh | rahul@customer.com |\n| Kavita Nair | kavita@customer.com |\n\n### Staff / Admin — web only (`/staff/login`)\n| Role | Email | After login |\n|------|-------|-------------|\n| Super Admin | superadmin@hotel.com | `/admin/dashboard` |\n| Admin Staff | staff@hotel.com | `/admin/dashboard` |\n\n### Vendor — web only (`/staff/login`)\n| Name | Email | After login |\n|------|-------|-------------|\n| Rajesh Kumar | vendor1@hotel.com | `/vendor/dashboard` |\n| Priya Sharma | vendor2@hotel.com | `/vendor/dashboard` |\n| Anil D'Souza | vendor3@hotel.com | `/vendor/dashboard` |",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/user/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new customer",
                "operationId": "125c7c1d419d843b818311eddaed12b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Amit Verma"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+919900000001"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered successfully"
                    }
                }
            }
        },
        "/user/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Customer login",
                "description": "Dummy accounts (password `password123`): amit@customer.com, sneha@customer.com, rahul@customer.com, kavita@customer.com",
                "operationId": "d1d381d6f4bfc8cb87925200f10817bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "amit@customer.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "mobile-app"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    }
                }
            }
        },
        "/user/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout (revoke current token)",
                "operationId": "6d8eecb1504d86ba23b75b4b9d0e6a82",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get authenticated customer",
                "operationId": "910e89184c62c5aa3322c6af36d27253",
                "responses": {
                    "200": {
                        "description": "Customer profile"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use token from POST /user/auth/login or /user/auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}